requirements.txt
pykeepass==4.0.6
GitPython==3.1.41
azure-devops==7.1.0b4
questionary==2.0.1